Basic history1

+++ src/orderinginfo.h
@@ -14,14 +14,14 @@ class OrderingInfo {
  public:
   /**
    * @brief Construct a new OrderingInfo
-   * 
+   *
    */
   OrderingInfo();
 
   /**
-   * @brief Increment the history heuristic value of the board for 
+   * @brief Increment the history heuristic value of the board for
    * the given color, from square, to square and depth.
-   * 
+   *
    * @param color Color to increment history for
    * @param from From square to increment history for
    * @param to To square to increment history for
@@ -30,9 +30,9 @@ class OrderingInfo {
   void incrementHistory(Color, int, int, int);
 
   /**
-   * @brief Lower history heuristic value of the board for 
+   * @brief Lower history heuristic value of the board for
    * the given color, from square, to square and depth.
-   * 
+   *
    * @param color Color to increment history for
    * @param from From square to increment history for
    * @param to To square to increment history for
@@ -42,26 +42,26 @@ class OrderingInfo {
 
   /**
    * @brief Update countermove.
-   * 
-   * @param color Color of the current sideToMove 
+   *
+   * @param color Color of the current sideToMove
    * @param counteredMove Previous move made in the Search
    * @param counterMove Move that rejected counteredMove
-   * 
+   *
    */
   void updateCounterMove(Color, int , int);
 
   /**
    * @brief Gets CounterMove using last move made in the position as info
-   * 
+   *
    * @param color Color of the current sideToMove
    * @param pMove Previous move made in the Search
    */
-  int getCounterMoveINT(Color, int) const; 
+  int getCounterMoveINT(Color, int) const;
 
   /**
    * @brief Get beta-cutoff history information for the given color, from square and
    * to square.
-   * 
+   *
    * @param color Color of side moving
    * @param from From square to get history for
    * @param to To square to get history for
@@ -71,7 +71,7 @@ class OrderingInfo {
 
   /**
    * @brief Update the killer moves for the given ply with the given move.
-   * 
+   *
    * @param ply Ply to update killer moves for
    * @param move Move to update killer moves with
    */
@@ -79,7 +79,7 @@ class OrderingInfo {
 
   /**
    * @brief Get the first killer move for the given ply.
-   * 
+   *
    * @param ply Ply to get killer move for
    * @return First killer move at the given ply
    */
@@ -87,7 +87,7 @@ class OrderingInfo {
 
   /**
    * @brief Get the second killer move for the given ply.
-   * 
+   *
    * @param ply Ply to get killer move for
    * @return Move Second killer move at the given ply
    */
@@ -113,14 +113,14 @@ class OrderingInfo {
   /**
    * @brief Table of beta-cutoff history values indexed by [color][from_square][to_square]
    */
-  int _history[2][64][64];
+  int16_t _history[2][64][64];
 
   /**
    * @brief Array of the moves (represented by their INT), that counter move
    * made on the previous ply (cause beta - cutoff)
    * Indexed by [OppositeColor][PieceType][to_square] of the move they countered
    */
-  int _counterMove[2][6][64]; 
+  int _counterMove[2][6][64];
 };
 
 #endif
